instagram/requirements.txt
asgiref==3.4.1
crispy-bootstrap5==0.6
crispy-tailwind==0.5.0
Django==3.2.7
django-adminlte-3==0.1.6
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-redis==5.0.0
djangorestframework==3.12.4
future==0.18.2
Markdown==3.3.4
Pillow==8.4.0
psycopg2-binary==2.9.1
pytz==2021.3
redis==3.5.3
Serializer==0.2.1
simplejson==3.17.5
sqlparse==0.4.2
